+++ src/analytics.ts
@@ -21,7 +21,7 @@ export const Registry = {
 };
 
 class AnalyticsPlatform {
-    private static nodeJsPlatformMapping = {
+    private static nodeJsPlatformMapping: Record<NodeJS.Platform, string> = {
         aix: 'desktop',
         android: 'android',
         darwin: 'mac',
@@ -31,9 +31,11 @@ class AnalyticsPlatform {
         sunos: 'desktop',
         win32: 'windows',
         cygwin: 'windows',
+        haiku: 'unknown',
+        netbsd: 'unknown',
     };
 
-    static for(p: string): string {
+    static for(p: NodeJS.Platform): string {
         return this.nodeJsPlatformMapping[p] || 'unknown';
     }
 }
@@ -107,8 +109,13 @@ export async function issueCommentEvent(site: DetailedSiteInfo): Promise<TrackEv
     return instanceTrackEvent(site, 'created', 'issueComment');
 }
 
-export async function issueWorkStartedEvent(site: DetailedSiteInfo): Promise<TrackEvent> {
-    return instanceTrackEvent(site, 'workStarted', 'issue');
+export async function issueWorkStartedEvent(
+    site: DetailedSiteInfo,
+    pushBranchToRemoteChecked: boolean,
+): Promise<TrackEvent> {
+    const attributesObject = instanceType({}, site);
+    attributesObject.attributes.pushBranchToRemoteChecked = pushBranchToRemoteChecked;
+    return instanceTrackEvent(site, 'workStarted', 'issue', attributesObject);
 }
 
 export async function issueUpdatedEvent(
@@ -164,7 +171,7 @@ export async function prCommentEvent(site: DetailedSiteInfo): Promise<TrackEvent
 }
 
 export async function prTaskEvent(site: DetailedSiteInfo, source: string): Promise<TrackEvent> {
-    const attributesObject: any = instanceType({}, site);
+    const attributesObject = instanceType({}, site);
     attributesObject.attributes.source = source;
     return trackEvent('created', 'pullRequestComment', attributesObject);
 }
@@ -657,33 +664,27 @@ function tenantOrNull<T>(e: Object, tenantId?: string): T {
     return newObj as T;
 }
 
-function instanceType(eventProps: Object, site?: DetailedSiteInfo, product?: Product): Object {
-    let attrs: Object | undefined = undefined;
-    const newObj = eventProps;
-
+function instanceType(
+    eventProps: Record<string, any>,
+    site?: DetailedSiteInfo,
+    product?: Product,
+): Record<string, any> {
     if (product) {
-        attrs = { hostProduct: product.name };
+        eventProps.attributes = eventProps.attributes || {};
+        eventProps.attributes.hostProduct = product.name;
     }
 
     if (site && !isEmptySiteInfo(site)) {
-        const instanceType: string = site.isCloud ? 'cloud' : 'server';
-        attrs = { instanceType: instanceType, hostProduct: site.product.name };
+        eventProps.attributes = eventProps.attributes || {};
+        eventProps.attributes.instanceType = site.isCloud ? 'cloud' : 'server';
+        eventProps.attributes.hostProduct = site.product.name;
     }
 
-    if (attrs) {
-        newObj['attributes'] = { ...newObj['attributes'], ...attrs };
-    }
-
-    return newObj;
+    return eventProps;
 }
 
-function excludeFromActivity(eventProps: Object): Object {
-    const newObj = eventProps;
-
-    if (newObj['attributes']) {
-        newObj['attributes'] = { ...newObj['attributes'], ...{ excludeFromActivity: true } };
-    } else {
-        Object.assign(newObj, { attributes: { excludeFromActivity: true } });
-    }
-    return newObj;
+function excludeFromActivity(eventProps: Record<string, any>): Record<string, any> {
+    eventProps.attributes = eventProps.attributes || {};
+    eventProps.attributes.excludeFromActivity = true;
+    return eventProps;
 }