+++ src/react/atlascode/startwork/startWorkController.ts
@@ -30,6 +30,7 @@ export interface StartWorkControllerApi {
         sourceBranch: Branch,
         targetBranch: string,
         upstream: string,
+        pushBranchToRemote: boolean,
     ) => Promise<{ transistionStatus?: string; branch?: string; upstream?: string }>;
     closePage: () => void;
     openJiraIssue: () => void;
@@ -122,6 +123,7 @@ export function useStartWorkController(): [StartWorkState, StartWorkControllerAp
             sourceBranch: Branch,
             targetBranch: string,
             upstream: string,
+            pushBranchToRemote: boolean,
         ): Promise<StartWorkResponseMessage> => {
             return new Promise<StartWorkResponseMessage>((resolve, reject) => {
                 (async () => {
@@ -136,6 +138,7 @@ export function useStartWorkController(): [StartWorkState, StartWorkControllerAp
                                 sourceBranch,
                                 targetBranch,
                                 upstream,
+                                pushBranchToRemote,
                             },
                             StartWorkMessageType.StartWorkResponse,
                             ConnectionTimeout,