+++ src/webviews/startWorkOnBitbucketIssueWebview.ts
@@ -105,6 +105,7 @@ export class StartWorkOnBitbucketIssueWebview
                                     e.targetBranchName,
                                     e.sourceBranch,
                                     e.remoteName,
+                                    e.pushBranchToRemote,
                                 );
                             }
 
@@ -138,6 +139,7 @@ export class StartWorkOnBitbucketIssueWebview
         destBranch: string,
         sourceBranch: Branch,
         remote: string,
+        pushBranchToRemote: boolean,
     ): Promise<void> {
         // checkout if a branch exists already
         try {
@@ -160,7 +162,11 @@ export class StartWorkOnBitbucketIssueWebview
             true,
             `${sourceBranch.type === RefType.RemoteHead ? 'remotes/' : ''}${sourceBranch.name}`,
         );
-        await repo.push(remote, destBranch, true);
+
+        if (pushBranchToRemote) {
+            await repo.push(remote, destBranch, true);
+        }
+
         return;
     }